+++ oauth-core/src/main/java/org/eclipse/lyo/server/oauth/core/OAuthRequest.java
@@ -13,9 +13,16 @@ package org.eclipse.lyo.server.oauth.core;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
+import java.util.Collections;
+import java.util.Enumeration;
+import java.util.HashMap;
+import java.util.Map;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletRequestWrapper;
+import javax.ws.rs.container.ContainerRequestContext;
+import javax.ws.rs.core.MultivaluedMap;
 
 import net.oauth.OAuth;
 import net.oauth.OAuthAccessor;
@@ -25,6 +32,7 @@ import net.oauth.OAuthProblemException;
 import net.oauth.OAuthValidator;
 import net.oauth.server.OAuthServlet;
 
+import org.eclipse.lyo.oslc4j.core.OSLC4JUtils;
 import org.eclipse.lyo.server.oauth.core.consumer.LyoOAuthConsumer;
 
 /**
@@ -52,11 +60,29 @@ public class OAuthRequest {
 	private HttpServletRequest httpRequest;
 	private OAuthMessage message;
 	private OAuthAccessor accessor;
-	
+
+	/**
+	 * Use {@link #OAuthRequest(HttpServletRequest, boolean)} instead.
+	 */
+	@Deprecated
 	public OAuthRequest(HttpServletRequest request)
 			throws OAuthException, IOException {
+		// preserving legacy behaviour
+		this(request, false);
+	}
+
+	/**
+	 * @param detectLyoURI Determines if {@link OSLC4JUtils#resolveFullUri(HttpServletRequest)} shall
+	 *                     be used.
+	 * @throws OAuthException OAuth request is invalid
+	 */
+	public OAuthRequest(HttpServletRequest request, boolean detectLyoURI) throws OAuthException, IOException {
 		this.httpRequest = request;
-		this.message = OAuthServlet.getMessage(httpRequest, null);
+		String fullUrl = null;
+		if (detectLyoURI) {
+			fullUrl = OSLC4JUtils.resolveFullUri(request);
+		}
+		this.message = OAuthServlet.getMessage(httpRequest, fullUrl);
 
 		LyoOAuthConsumer consumer = OAuthConfiguration.getInstance()
 				.getConsumerStore().getConsumer(message);
@@ -71,15 +97,62 @@ public class OAuthRequest {
 		String token = this.message.getToken();
 		if (token != null) {
 			this.accessor.tokenSecret = OAuthConfiguration.getInstance()
-					.getTokenStrategy().getTokenSecret(this.httpRequest, token);
+					.getJaxTokenStrategy().getTokenSecret(token);
 		}
 	}
-	
+
+	public static class OAuthServletRequestWrapper extends HttpServletRequestWrapper {
+
+		private final Map<String, String[]> formParams;
+
+		/**
+		 * Constructs a request object wrapping the given request.
+		 *
+		 * @param request
+		 * @throws IllegalArgumentException if the request is null
+		 */
+		public OAuthServletRequestWrapper(HttpServletRequest request,
+										  MultivaluedMap<String, String> formParams) {
+			super(request);
+			this.formParams = aggregateMultimap(formParams);
+		}
+
+		private Map<String, String[]> aggregateMultimap(MultivaluedMap<String, String> multimap) {
+			HashMap<String, String[]> map = new HashMap<>();
+			multimap.forEach((key, strings) -> map.put(key, strings.toArray(new String[0])));
+			return map;
+		}
+
+		@Override
+		public String getParameter(String name) {
+			String[] values = formParams.get(name);
+			if (values == null || values.length == 0) {
+				return null;
+			}
+			return values[0];
+		}
+
+		@Override
+		public Map<String, String[]> getParameterMap() {
+			return formParams;
+		}
+
+		@Override
+		public Enumeration<String> getParameterNames() {
+			return Collections.enumeration(formParams.keySet());
+		}
+
+		@Override
+		public String[] getParameterValues(String name) {
+			return formParams.get(name);
+		}
+	}
+
 	public HttpServletRequest getHttpRequest() {
 		return httpRequest;
 	}
 
-	public void setHttpRequest(HttpServletRequest httpRequest) {
+	private void setHttpRequest(HttpServletRequest httpRequest) {
 		this.httpRequest = httpRequest;
 	}
 
@@ -112,7 +185,7 @@ public class OAuthRequest {
 		try {
 			OAuthConfiguration config = OAuthConfiguration.getInstance();
 			config.getValidator().validateMessage(message, accessor);
-			config.getTokenStrategy().validateAccessToken(this);
+			config.getJaxTokenStrategy().validateAccessToken(this);
 		} catch (URISyntaxException e) {
 			throw new ServletException(e);
 		}