+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## 4.0.0 (WIP)
 
+### Added
+
+- `OAuthRequest` can now detect configured server URI via `OSLC4JUtils`
+
 ### Changed
 
 - `HTTPConstants` is no longer a public class
 - `ServletListener` was renamed to `OAuthServletListener`
-